# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cinema API",
        "description": "API for Cinema",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Localhost API Server"
        }
    ],
    "paths": {
...
        "/api/seances": {
            "get": {
                "tags": [
                    "Séances"
                ],
                "summary": "Liste toutes les séances",
                "description": "Récupère la liste de toutes les séances de cinéma.",
                "operationId": "eb8af953bc7b9646b2294583ed472f06",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "film_id": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "salle_id": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "start_time": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-04-01T14:00:00Z"
                                            },
                                            "session": {
                                                "type": "string",
                                                "example": "Soir"
                                            },
                                            "langue": {
                                                "type": "string",
                                                "example": "Français"
                                            },
                                            "type_seance": {
                                                "type": "string",
                                                "example": "VIP"
                                            },
                                            "prix": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 12.5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Séances"
                ],
                "summary": "Créer une nouvelle séance",
                "description": "Ajoute une nouvelle séance de cinéma.",
                "operationId": "c4cf111697bf9bcbe40078debd8842f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "film_id",
                                    "salle_id",
                                    "start_time",
                                    "session",
                                    "langue",
                                    "type_seance",
                                    "prix"
                                ],
                                "properties": {
                                    "film_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "salle_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-04-01T14:00:00Z"
                                    },
                                    "session": {
                                        "type": "string",
                                        "example": "Soir"
                                    },
                                    "langue": {
                                        "type": "string",
                                        "example": "Français"
                                    },
                                    "type_seance": {
                                        "type": "string",
                                        "enum": [
                                            "Normale",
                                            "VIP"
                                        ],
                                        "example": "VIP"
                                    },
                                    "prix": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 12.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Séance créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "film_id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "salle_id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "start_time": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-04-01T14:00:00Z"
                                        },
                                        "session": {
                                            "type": "string",
                                            "example": "Soir"
                                        },
                                        "langue": {
                                            "type": "string",
                                            "example": "Français"
                                        },
                                        "type_seance": {
                                            "type": "string",
                                            "example": "VIP"
                                        },
                                        "prix": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 12.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/seances/type/{type}": {
            "get": {
                "tags": [
                    "Séances"
                ],
                "summary": "Filtrer les séances par type",
                "description": "Récupère les séances en fonction de leur type (Normale ou VIP).",
                "operationId": "c85ed350dd9cabb12aa45f8e612530be",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Type de séance (Normale ou VIP)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "400": {
                        "description": "Type invalide"
                    },
                    "404": {
                        "description": "Aucune séance trouvée"
                    }
                }
            }
        },
...
    },
...
}